Assignment_3_ArmstrongNumbers.py added to repo
Task:
Find out if a given number is an "Armstrong Number".
An n-digit number that is the sum of the nth powers of its digits is called an n-Armstrong number. Example : 371 = 33 + 73 + 13; 9474 = 94 + 44 + 74 + 44.
Write a Python program that;
takes a positive integer number from the user,
checks the entered number if it is Armstrong,
consider the negative, float and any entries other than numeric values then display a warning message to the user.